+++ gui/src/renderer/transitions.ts
@@ -45,6 +45,7 @@ const transitionRules = [
   r('/settings', '/settings/preferences', transitions.push),
   r('/settings', '/settings/advanced', transitions.push),
   r('/settings/advanced', '/settings/advanced/wireguard-keys', transitions.push),
+  r('/settings/advanced', '/settings/advanced/linux-split-tunneling', transitions.push),
   r('/settings', '/settings/support', transitions.push),
   r(null, '/settings', transitions.slide),
   r(null, '/select-location', transitions.slide),