+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Line wrap the file at 100 chars.                                              Th
 - Add fish shell completions for the mullvad CLI.
 - Reconnect with a new key when WireGuard key is rotated automatically, previously the tunnel would
   time out before reconnecting.
+- Add split tunneling menu under advanced settings in Linux app.
 
 #### Android
 - Add split-tunnelling, allowing apps to be configured to be excluded from the tunnel.